requirements.txt
imageio==2.22.0
matplotlib==3.5.2
numpy==1.22.3
Pillow==9.2.0
tqdm==4.64.0
